Allow more control over Rust layout.
Boxing newtype structs in enum variants is now allowed to work even if
it results in a slightly different layout from the OCaml.  This allows
us to use fewer Rcs in many cases and significantly reduce the size of a
Constr.  It's not clear whether this will ultimately yield savings when
we switch to arenas (since it may turn out to be hard to reliably get the
same level of sharing with a different layout, at least without doing
more work up front), but it's nice to have options.