Experiment: trade off indirection for cheap drop.
Instead of directly referencing ORef<T>, instead reference &'b ORef<T>.
This way, the code mostly goes through as it did before, but we avoid
actually incrementing the reference count.  This avoids the shared write
and hopefully makes the drop smoother when the arena is destroyed.
Before doing this, I checked to see whether this would cause any
annoyances down the line; the biggest one seems to be that we do
actually create a new Projection sometimes when evaluating FTerms
(during eta_expand_ind_stack).  However, it looks like something that
should be easy to work around.

It's worth noting that the indirection we see here isn't really
fundamental.  We could get just as little indirection as with the Rc by
pointing directly into it.  There isn't that much downside, except that
it prevents us from sharing existing Rc'd terms when we go back to
Constr from FConstr; however, that may end up being a small price to pay
in most cases given how many times we expect to allocate FTerms vs. the
number of conversions back to Constr we do.  For now, we just keep it
this way; it seems likely we'll want to move to something with less
indirection eventually.