Added part of iota reduction (arg extraction).
Also added a bit more type safety by restricting certain functions that
expect to act on stacks with just Shift + App to such stacks; as a
result, we can avoid some assertions about stack well-formedness.
However, this cute method may have to be abandoned later if we switch to
a more efficient representation of stacks.