Added knh and knht.
These are the machines that inspect the head of a term until they find a
potential redex or neutral term.  There is some silly duplication
between knh and knht that can probably be eliminated in any number of
nice ways.

This is (hopefully) the last step before knr, the function that computes
the weak-head normal form of a term.