Parse Maps as HashMaps.
We could also choose BTreeMaps, if we wanted (and/or were worried about
key duplication), since they are closer to the actual implementation of
Map in OCaml.  This was sort of planned already, but it turns out to be
useful for making knr work since the final step requires performing map
operations on deserialized Map structures, and it would be irritating to
work with the OCaml version unless we had a good reason.