Adding remaining public methods from closure.
This also included actually creating the Infos table, which led to
rejiggering lifetimes in an effort to pave the way for implementing
reductions properly.  The main change is that Globals is now accessed
directly from Infos, rather than indirectly through Env, so the
lifetimes are a bit simpler and we can split up Globals into different
parts (which may be necessary for reduction to have access to all the
information it might need, since it seems to need a few things from
engagement).