Add missing prerequisites to closure definitions.
All of them concern the necessity of typechecking a parameter before
calling the function.  It really would be nice to just incorporate that
into Rust's type system (or alternatively, scrap the requirement and
perform the relevant checks in code, but then we would still not have
some nice properties like termination of evaluation).