Adding the first few reduction functions.
Unfortunately, we also had to (temporarily) abandon SmallVec; the
primary reason for this is that SmallVec's Drop implementations don't
implement #[may_dangle].  However, I don't think there's any fundamental
reason why they can't do so (given that Vec does); it's just a matter of
putting in the work to make this safe for SmallVec (though I suspect
it's a bit more complicated than just adding it, since SmallVec may not
use PhantomData appropriately etc.).  Also, we'd have to figure out how
to enable the flag for people who don't care about stable Rust, since
the #[may_dangle] flag isn't stable (for good reasons, probably).

Worth noting that if we end up putting Subs instances in arenas (which
would perhaps be a good idea anyway if we use SmallVecs, given that the
size of FTerms would increase), we would probably not run into this
problem anymore.  So we might never have to do the above; but it would
still be a good idea.