Finished conversion (reductions.ml).
This was mostly straightforward, but required implementing a bunch of
explicit equality functions (which I'd been avoiding).  I also fixed a
few places where we were relying on Rust's #[derive(PartialEq,Eq)]
version of equality, which isn't quite right for Coq in some cases.
There are still probably more places that rely on Rust's equality (or
hash function) that have to be fixed in this manner.