The CMap -> Map conversion was premature.
Enabling this caused huge size bloat in the deserialization of some .vo
files (20x+ size blowup).  I am not sure whether this is because Coq
actually utilizes the persistence of the maps significantly, or just
some artifact of the way the map conversion was done, but clearly we
need a different approach.

Also worth noting that there does appear to be *some* sharing of the
structures we shared before (the ones that we weren't sure about, not
the ones that were actually checked), but avoiding it doesn't have
nearly the size overhead of avoiding sharing Maps.  That being said, we
should keep an eye on this because the sharing for Sfbs etc. might prove
more important to typecheckinga and reduction performance.