Fix looping reduction problem in Coq.
In the current implementation of Coq, there is a loop during reduction
that tries to deal with situations where reduction of one side of a
pair of terms affects the other one, even if it's already in weak-head
normal form.  This can cause spurious reduction failures, so the
existing solution is to just loop and try again if the first side isn't
in whnf after both are reduced; however, there's no proof that this
strategy must always terminate, and it's very ad hoc.

Here, we take advantage of the additional ownership and mutability
information that Rust provides to eliminate this possibility, by making
sure that the two terms have separate resources during reduction (with a
few benign exceptions, which only change monotonically in uninteresting
ways).  Therefore, we can get rid of the loop and the check to make sure
both sides are still in whnf.