Make FConstr and all related terms Send.
From here, parallelizing conversion should be quite easy :)

The most eyebrow-raising part is the first addition of unsafe code to
the checker; this is the new `GhostCell` abstraction, which allows one
to (roughly speaking) reify ownership of a large group of elements into
a single, logical owner, called a Set.  In order to access any element
in a Set at a particular mutability and lifetime, one must take out the
same mutability and lifetime on the owner, which is the safety argument
here.  Using this technique, we can mutably modify the interior of
FTerms without using atomics, locks, or Cells, as long as we have
posession of the owner; at the same time, the FTerms (and the owner)
remain Send and Sync and can be shared freely between threads and passed
through channels.  This is critical because it allows us to avoid
expensive synchronization during the bulk of conversion--synchronization
is only required when threads notify each other that they are finished
and move ownership (and a small amount of other, necessary information)
around.  It is worth noting that the ownership is entirely at compile
time--there is no additional overhead whatsoever compared to a Cell
(there can actually be a little less since you don't need to copy things
out of the Cell, though we generally do just copy things out of the
Cell in this case).

The other things that were addressed were more pedestrian: we can't take
Globals mutably anymore if we want to share it between the threads
during reduction, so we replaced the mutable ownership and custom lazy
initialization with an efficient synchronized lazy initializer (the
synchronization should not matter much in practice because [1] usually
forcing doesn't have to do anything, and [2] we cache globals really
quickly once we see them during reduction).  We also, sadly, changed all
the Rcs in Constrs to Arcs; as yet, there's still no obvious clever way
to tell Rust "we aren't going to clone these Rcs anymore" (and in fact,
we aren't respecting any such invariant right now!).  Performance of
loading the file has almost certainly regressed somewhat, but I'm not
sure how much; in any case, it's still small enough that it shouldn't
swallow up other improvements from parallelism (I hope).  Fortunately,
during reduction and conversion we usually take Constrs by reference
rather than cloning them.

The next step is hooking in Rayon or Crossbeam.