Adding universe substituitions.
This *really* completes reductions (except for module substitutions,
which we defer to later).

Also sort of adds hash consing, or at least hash consing infrastructure.
It's not robust, and doesn't actually do any hash consing, but that's
completely okay because (1) it's not relied upon for soundness, and
(2) the next release of Coq gets rid of hash consing in the checker
(at least for universes).